# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service Verifikasi User API",
        "description": "Dokumentasi interaktif untuk Tugas 2 IAE - Service Verifikasi",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/verifications": {
            "get": {
                "tags": [
                    "Verifications"
                ],
                "summary": "Mengambil daftar semua pengajuan verifikasi",
                "operationId": "30ce2306a3c84ed3ed06788ccbb3d1b0",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar verifikasi"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Verifications"
                ],
                "summary": "Mengirim data pengajuan verifikasi baru",
                "operationId": "d87f40d2e307f3797683d2387a515a97",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "nik",
                                    "bank_account_number"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nik": {
                                        "type": "string",
                                        "example": "3271234567890001"
                                    },
                                    "bank_account_number": {
                                        "type": "string",
                                        "example": "1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data verifikasi berhasil dikirim"
                    },
                    "400": {
                        "description": "Validasi Gagal"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/verifications/{id}": {
            "get": {
                "tags": [
                    "Verifications"
                ],
                "summary": "Mengecek status verifikasi berdasarkan user_id",
                "operationId": "f6bbd93e0b06d407ee4b758b3c4634c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID User",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail verifikasi"
                    },
                    "404": {
                        "description": "Data verifikasi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Verifications"
                ],
                "summary": "Memperbarui status verifikasi",
                "operationId": "f774ac78283cdce7969344beb888952d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Verification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "verification_status"
                                ],
                                "properties": {
                                    "verification_status": {
                                        "type": "string",
                                        "example": "VERIFIED",
                                        "enum": [
                                            "VERIFIED",
                                            "NOT_VERIFIED"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status verifikasi berhasil diperbarui"
                    },
                    "400": {
                        "description": "Validasi Gagal"
                    },
                    "404": {
                        "description": "Data verifikasi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Verifications",
            "description": "Verifications"
        }
    ]
}